+++ permissions/src/epilot/roles.py
@@ -133,7 +133,7 @@ class Roles:
         
         url = utils.generate_url(operations.PutRoleRequest, base_url, '/v1/permissions/roles/{roleId}', request)
         headers = {}
-        req_content_type, data, form = utils.serialize_request_body(request, "role_payload", True, True, 'json')
+        req_content_type, data, form = utils.serialize_request_body(request, "role_payload", False, True, 'json')
         if req_content_type not in ('multipart/form-data', 'multipart/mixed'):
             headers['content-type'] = req_content_type
         headers['Accept'] = 'application/json'